Enlarge maximum payload for secret to 10100 chars (and 101000 bytes total querysize)
While trying to exchange a signed JWT with a collegue, we noticed, that
the secret was silently truncated. Given that the signed JWT is basicly
base64 encoded data it is not directly obvious that the data was
incomplete.

This change enlarges the payload into a range where it can be used to
transfer RSA keys, JWT and similar material, which also must be kept
confidential.

The change also adds an output for the user, that reflects the field
limit and the used chars.